fixing attendance mangement
